% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendro}
\alias{dendro}
\title{Hierarchical Cluster (Dendrogram)}
\usage{
dendro(veg, method = "bray", env = NULL, group.col = NULL,
  colvec = "black", pch = 20, cex = 1, lab.cex = 0.7)
}
\arguments{
\item{veg}{vegetation matrix}

\item{method}{method of the dissimilarity index}

\item{env}{environment matrix with the group information}

\item{group.col}{column in which the group is located}

\item{colvec}{colour vector of the groups; like c("col1","col2","col3") or c(1,2,3)}

\item{pch}{point vector of the groups; like c("shape1","shape2","chape3") or c(1,2,3)}

\item{cex}{scaled plot}

\item{lab.cex}{scale of label}
}
\description{
Creates a hierachical dendrogram. It is also possible to color the leaves by group.
}
\note{
Information about the different dissimilarity indexes read \code{\link[vegan]{vegdist}}
}
\examples{
library(vegan)
data(dune)
data(dune.env)

# Simple dendrogram
den <- dendro(veg = dune)
plot(den, horiz = TRUE)

# Dendrogram coloured by group
Cvec = c("black", "gray20", "gray40", "gray80")
Pch = c(15, 16, 17, 18)

den <- dendro(veg = dune, env = dune.env, group.col = 3, colvec = Cvec, pch = Pch, cex = 2)
plot(den, leaflab = "none", horiz = TRUE)
 
}
