% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_boxplot_2G.R
\name{env_boxplot_2G}
\alias{env_boxplot_2G}
\title{Produces boxplots for than two groups}
\usage{
env_boxplot_2G(env, type = NULL, group.col = 0)
}
\arguments{
\item{env}{table of data (must be numeric)}

\item{type}{selection of test (“u.test”, “t.test”)}

\item{group.col}{column of the group info (must be a factor)}
}
\description{
Produces boxplots and performs a U-Test or a t-Test.
}
\note{
Information about the U-Test read \code{\link[stats]{wilcox.test}} and the t-Test read \code{\link[stats]{t.test}}
}
\examples{
dat <- iris[1:100,]

results <- env_boxplot_2G(env = dat, type = "u.test", group.col = 5)
results

}
