% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DCA.R
\name{plot_DCA}
\alias{plot_DCA}
\title{Plots a Detrended correspondence analysis}
\usage{
plot_DCA(veg, env, group.col = 0, r.cutoff_env = 0.3,
  r.cutoff_spec = 0.3, colvec, pch = 20, lty = 1, lwd = 1,
  cex = 1, cex.lab = 1, cex.leg = 1, ordihull = F)
}
\arguments{
\item{veg}{vegetation matirx}

\item{env}{envirement matrix}

\item{group.col}{collom in which the group is standing}

\item{r.cutoff_env}{the R²-cutoff of the enviremnt}

\item{r.cutoff_spec}{the R²-cutoff of the species}

\item{colvec}{colorvector of the groups}

\item{pch}{point shape}

\item{lty}{line shape}

\item{lwd}{line size}

\item{cex}{scaled plot}

\item{cex.lab}{scale of label}

\item{cex.leg}{scale of legend}

\item{ordhull}{circled the groups}
}
\value{
Plots a grouped DCA with fitted species and envirment vectors
}
\description{
Pointts are the sites
}
\examples{
out <- select.envfit(select.envfit(envfit(DCA, Env, perm = 1000), r.select = "0.3"))

}
