% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.envfit.R
\name{select.envfit}
\alias{select.envfit}
\title{Dose the R²-cutoff}
\usage{
select.envfit(fit, r.select)
}
\arguments{
\item{fit}{fitted vecoter from vegan}

\item{r.select}{r²-cutoff}
}
\value{
the fitted vectors without the r² value higher than the cutoff
}
\description{
Selects the fitted vectors of a ordination by a certai the r²-cutoff
}
\examples{
fit <- envfit(ord, env, permutations = 1000)
out <- select.envfit(fit, r.select = 0.3))
out

}
