% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_trans.R
\name{veg_trans}
\alias{veg_trans}
\title{Transforming expended Braun-Blanquet-Categories}
\usage{
veg_trans(veg)
}
\arguments{
\item{veg}{vegetation data frame}
}
\description{
Transforms the expended Braun-Blanquet-Categories of a vegetation data frame
into percentage
}
\note{
Changes the Braun-Blanquet-Categories in to following proportions: \cr
r --> 0.01\% \cr
 + --> 0.5\% \cr
 1 --> 2.5\% \cr
 2a --> 8.8\% \cr
 2b --> 20\% \cr
 3 --> 37.5\% \cr
 4 --> 62.5\% \cr
 5 --> 87.5\% \cr
}
\examples{
NOT RUN

}
