% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.envfit.R
\name{select.envfit}
\alias{select.envfit}
\title{Dose the R²-cutoff}
\usage{
select.envfit(fit, r.select)
}
\arguments{
\item{fit}{fitted enviermental vecoter from vegen}

\item{r.select}{R²-cutoff}
}
\value{
the fittet enviremantal vectors witour the r² valu ounder the cutoff
}
\description{
Selects the fitted vectors of a ordination by the R²-cutoff
}
\examples{
fit <- envfit(ord, env, permutations = 1000)
out <- select.envfit(fit, r.select = 0.3))
out

}
